backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
requests>=2.31.0

# ML / AI (CPU optimized for Render)
torch==2.9.0+cpu
torchvision==0.24.0+cpu
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
h5py>=3.9.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scipy>=1.11.0
xarray>=2023.1.0
netcdf4>=1.6.0

# Database (Neon PostgreSQL)
psycopg2-binary>=2.9.0